Getting gold!
I got bored and decided to put some work in on the wumpus() function. Before it just picked a random direction to shoot in which led to death a lot of the times. Now, based on the direction that we have been traveling there will only ever be two slots that are possible for the wumpus to be in, so we shoot in those directons, and keep moving.

We are now getting gold! I made it so that if we see 'G' in the precepts it just returns 0 because I wanted to see if we were actually ever finding it, which we are!

Now we have to grab it, and get out. The process of grabbing it will probably be like the wumpus() function where we grab in all possible directions, but i'm not 100% on that.

I also have to handle some instances where we are experiencing infinite loops in the wumpus() function, but that shouldn't be hard. If we are ever experiencing the symptoms of an infinite loop we can just choose a random direction to go in until the symptoms are gone (in case you were wondering how i've been doing it so far).