More Stable - 500 Games no infinite loop
I took out the return 0 in the grabGold function, so that we just keep on bouncing around in the cave.

I handled all (hopefuly all) of the infinite loops that we ran into, sometimes making random decisions based on the circumstances.

I ran a successful test of 200 games, and then 500 games, but I encourage you all to try to test some more (I am tired rn)

In order to move forward we have to change the grab gold function so that it grabs correctly, and then we need to call the escape function (which at that point needs to take priority over everything else)